fix: Modify the create table query and insert record queries
The previous version forgets to put PRIMARY KEY constrain and NOT
NULL constrain when creating tables and it cause some insert record
errors (inserting a duplicated record into a table). In this version
,it puts the adequate PK and NOT NULL constrain on each table to avoid
the previous issues. Apart from this, this version also make a change
on insert record queries. For the user table, it is important to change
the level column value when a user become a "paid" member from a "free"
member. To deal with this, this version put an "on conflict" statement
in each insert query and tell the system what it should do.