DOCS: Write docstring for each function and print() in main function
The previous version forgets to write the docstring for each function.
This make other developer hardly understand the code. This version fix
this issue. Also, this version puts print statement in main functuon.
It allows users to understand which stage this code is operating when
they run this code in terminal.